RgbCube/src/requirements.txt
torch==1.5.0
numpy==1.16.5
pptk==0.1.0
matplotlib==3.1.1
pandas==0.25.1
Pillow==8.0.1
scikit_learn==0.23.2
